// src/app/modules/realtor/model/realtor.manager.model.ts
export interface RealtorManagerModel {
	id: number;
	cellphone: string;
	email: string;
	lastName: string;
	licenseNumber: string;
	mailingAdd: string;
	name: string;
	notes: string;
	brokerCompanyId: number;
	brokerCompanyName: string;
	brokerCompanyPhone: string;
	brokerCompanyPhysicalAdd: string;
	brokerCompanyWebSite: string;
}